+++ cli/admin/help.go
@@ -54,3 +54,9 @@ func getFuncInfo(funcName string, conn *connector.Conn) (*FuncInfo, error) {
 	return &funcInfo, nil
 
 }
+
+var (
+	funcHelpMsgTmpl = `{{ .FuncInfo }}{{ if .ArgsUsage }}
+Args:
+{{ .ArgsUsage }}{{ end }}`
+)
