--- cli/admin/lua/func_help_msg_template.lua
@@ -1,4 +1,0 @@
-{{ .FuncInfo }}{{ if .ArgsUsage }}
-
-Args:
-{{ .ArgsUsage }}{{ end }}
