+++ cli/connector/lua/eval_func_template.lua
@@ -1,16 +1,16 @@
 local function func(...)
-	{{ .FunctionBody }}
+    {{ .FunctionBody }}
 end
 local args = require('msgpack').decode(string.fromhex('{{ .ArgsEncoded }}'))
 
 local ret = {
-	load(
-		'local func, args = ... return func(unpack(args))',
-		'@eval'
-	)(func, args)
+    load(
+        'local func, args = ... return func(unpack(args))',
+        '@eval'
+    )(func, args)
 }
 return {
-	data_enc = require('digest').base64_encode(
-		require('msgpack').encode(ret)
-	)
+    data_enc = require('digest').base64_encode(
+        require('msgpack').encode(ret)
+    )
 }
