+++ cli/repair/lua/reload_clusterwide_config_func_body.lua
@@ -8,7 +8,7 @@ assert(err == nil, string.format('Failed to load new config: %s', err))
 
 local current_uuid = box.info().uuid
 if cfg:get_readonly().topology.servers[current_uuid] == nil then
-	return false
+    return false
 end
 
 local roles_configured_state = 'RolesConfigured'
@@ -17,18 +17,18 @@ local connecting_fullmesh_state = 'ConnectingFullmesh'
 local state = confapplier.wish_state(roles_configured_state, wish_state_timeout)
 
 if state == connecting_fullmesh_state then
-	error(string.format(
-		'Failed to reach %s config state. Stuck in %s. ' ..
-			'Call "box.cfg({replication_connect_quorum = 0})" in instance console and try again',
-		roles_configured_state, state
-	))
+    error(string.format(
+        'Failed to reach %s config state. Stuck in %s. ' ..
+            'Call "box.cfg({replication_connect_quorum = 0})" in instance console and try again',
+        roles_configured_state, state
+    ))
 end
 
 if state ~= roles_configured_state then
-	error(string.format(
-		'Failed to reach %s config state. Stuck in %s',
-		roles_configured_state, state
-	))
+    error(string.format(
+        'Failed to reach %s config state. Stuck in %s',
+        roles_configured_state, state
+    ))
 end
 
 cfg:lock()
