+++ cli/replicasets/lua/edit_instance_body.lua
@@ -3,7 +3,7 @@ local cartridge = require('cartridge')
 local servers = ...
 
 local res, err = cartridge.admin_edit_topology({
-	servers = servers,
+    servers = servers,
 })
 
 assert(err == nil, tostring(err))
