--- cli/replicasets/lua/table_template.lua
@@ -1,1 +1,0 @@
-{ {{ .OptsString }} }
