+++ cli/replicasets/lua/topology_replicasets_body_template.lua
@@ -8,8 +8,8 @@ local replicasets, err = cartridge.admin_get_replicasets()
 assert(err == nil, tostring(err))
 
 for _, replicaset in pairs(replicasets) do
-	local topology_replicaset = {{ .FormatTopologyReplicasetFuncName }}(replicaset)
-	table.insert(topology_replicasets, topology_replicaset)
+    local topology_replicaset = {{ .FormatTopologyReplicasetFuncName }}(replicaset)
+    table.insert(topology_replicasets, topology_replicaset)
 end
 
 return unpack(topology_replicasets)
